// enums/applicants/applicant-voe-form.enum.ts
export enum ApplicantVoeFormEnum {
  EMPLOYED_BY_US = "EMPLOYED_BY_US",
  WAS_EMPLOYED_AS = "WAS_EMPLOYED_AS",
  DID_DRIVE_FOR_YOU = "DID_DRIVE_FOR_YOU",
  SAFETY_PERFORMANCE_HISTROY_REPORT = "SAFETY_PERFORMANCE_HISTROY_REPORT",
  REGISTERED_ACCIDENTS_DETAILS = "REGISTERED_ACCIDENTS_DETAILS",
  ACCIDENT_REPORTED_TO_GOVERNMENT = "ACCIDENT_REPORTED_TO_GOVERNMENT",
  REASON_TO_LEAVE_EMPLOYMENT = "REASON_TO_LEAVE_EMPLOYMENT",
  SIGNATURE_VOE = "SIGNATURE_VOE",
  SENDER_INFO = "SENDER_INFO",
}